hardware/requirements.txt
opencv-python
pyaudio
numpy
pyserial
smbus
wave